==> notebooks/0-SuburbMetadataClean.ipynb <==
# Suburb coordinate cleaning


import os
import numpy as np
import pandas as pd

from geopy.geocoders import GoogleV3  # for determining uni campus coordinates
from geopy.distance import geodesic  # for calculating distance between coordinates

# make all output interactive
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import helper

# Define file structure constants
# ================================
DATA_PATH = os.path.join(os.getcwd(), "../data/")
RAW_DATA_PATH = os.path.join(DATA_PATH, "raw")
DERIVED_DATA_PATH = os.path.join(DATA_PATH, "derived")
CORR_DATA_PATH = os.path.join(DATA_PATH, "correspondence")


## Coordinates

The data provided in [`australian_postcodes.csv`](../data/raw/australian_postcodes.csv) is fairly accurate; however, various discripancies have been noted in the coordinates for the suburbs. This data will be cleaned using `geopy` and the `GoogleV3` engine to determine the coordinates for each suburb.

In order for the `GoogleV3` code to work, VSCode must be opened from the terminal using the following code to ensure the environment variable for the Google Maps API key is set correctly and inherited by the session:

``` bash
> source ~/.bash-profile
> code .
```

geolocator = GoogleV3(api_key=os.getenv("GMAPS_APIKEY"))

We collect the 'raw' data using the `getSuburbsMetadata()` function defined in [`helper.py`](helper.py). Since we are recalculating the coordinates for each suburb, we only need the name and postcode (we also use postcode just to be safe).

suburb_data = helper.getSuburbsMetadata("raw")
suburb_details = suburb_data[["postcode", "locality"]].to_numpy()

# each coordinate in coordinates_list corresponds to each row in the suburb_data
# dataframe, and so can be directly "joined" to the dataframe
coordinates_details = []
failed_details = []
previous_message_length = 0

for index, details in enumerate(suburb_details):
    search_string = f"{details[1]} VIC, {details[0]}"
    location = geolocator.geocode(search_string)
    percentage_progress = round(((index + 1) / len(suburb_details)) * 100, 2)

    if location is not None:  # geocode succeeded -> store coordinates
        # message to user
        message = f"\r({str(percentage_progress).ljust(5,'0')}%) {str('[' + str(index + 1) + '/' + str(len(suburb_details)) + ']').ljust(15)}Geocoded: {details[1]}"
        print(" " * previous_message_length, end="\r")  # clear line
        print(message, end="\r")  # print message
        previous_message_length = len(message)  # record message length for next clear
        # record coordinate point
        point = location.point
        tup = (point.latitude, point.longitude)
        coordinates_details.append([details[0], details[1], tup])

    else:  # geocode failed -> record failure to remove suburb from dataset
        print(f"({str(percentage_progress).ljust(5,'0')}%) {str('[' + str(index + 1) + '/' + str(len(suburb_details)) + ']').ljust(15)}Failed: {details[1]}")
        failed_details.append(details)

In the process of adding our new `coordinates` column, we also remove the suburbs whose geocoding failed since we perform an inner join which, of course, is not completed for the suburbs whose coordinates weren't added to the `coordinates_details` array.

# convert array into dataframe
coordinates_df = pd.DataFrame(coordinates_details, columns=["postcode", "locality", "coordinates"])
# peform inner join
new_suburb_data = suburb_data.merge(coordinates_df[["locality", "coordinates"]], how="inner", on="locality")

new_suburb_data.shape
new_suburb_data.head()

print("Number of unique locality names:", len(new_suburb_data.locality.unique()))

## 2016 to 2021 conversion of ABS structure codes/names

All columns specifying the ABS structure codes and names are from 2016, so to match the new data (for showing details on WebApp) we also need to include the new codes and names from 2021.

This step also removes any suburbs from the raw suburbs dataset which are not considered in the ABS census statistics.

sa1_corr = pd.read_csv(os.path.join(CORR_DATA_PATH, "CG_SA1_2016_SA1_2021.csv"))
sa2_corr = pd.read_csv(os.path.join(CORR_DATA_PATH, "CG_SA2_2016_SA2_2021.csv"))
sa3_corr = pd.read_csv(os.path.join(CORR_DATA_PATH, "CG_SA3_2016_SA3_2021.csv"))
sa4_corr = pd.read_csv(os.path.join(CORR_DATA_PATH, "CG_SA4_2016_SA4_2021.csv"))

new_suburb_data = new_suburb_data\
    .merge(sa1_corr, on="SA1_MAINCODE_2016", suffixes=[None, "_SA1"])\
    .merge(sa2_corr, on="SA2_MAINCODE_2016", suffixes=[None, "_SA2"])\
    .merge(sa3_corr, on="SA3_CODE_2016", suffixes=[None, "_SA3"])\
    .merge(sa4_corr, on="SA4_CODE_2016", suffixes=[None, "_SA4"])

new_suburb_data.shape
new_suburb_data.head()

However, this has introduced some duplicate entries for suburbs. These are removed using the `drop_duplicates()` method on the `locality` column.

new_suburb_data = new_suburb_data.drop_duplicates("locality")
new_suburb_data.shape

## Remove unnecessary columns

Quite a few columns aren't necessary for our purposes, so these columns can be removed.

new_suburb_data.columns

new_suburb_data = new_suburb_data[[
    "postcode", "locality", "coordinates", "lgaregion", "SA1_MAINCODE_2016", 
    "SA1_CODE_2021", "SA2_MAINCODE_2016", "SA2_NAME_2016_SA2", "SA2_CODE_2021", 
    "SA2_NAME_2021", "SA3_CODE_2016", "SA3_NAME_2016_SA3", "SA3_CODE_2021", 
    "SA3_NAME_2021", "SA4_CODE_2016", "SA4_NAME_2016_SA4", "SA4_CODE_2021", 
    "SA4_NAME_2021"
]]
new_suburb_data.columns = [
    "postcode", "locality", "coordinates", "lgaregion", "sa1_code_2016",
    "sa1_code_2021", "sa2_code_2016", "sa2_name_2016", "sa2_code_2021",
    "sa2_name_2021", "sa3_code_2016", "sa3_name_2016", "sa3_code_2021",
    "sa3_name_2021", "sa4_code_2016", "sa4_name_2016", "sa4_code_2021",
    "sa4_name_2021"
]

new_suburb_data.shape
new_suburb_data.head()

## Remove 'bad' rows/suburbs

There are some suburbs whose geocoded coordinates are outside the Victorian state boundaries.

Victoria's longitude can range from 140.9–150, so any rows with coordinates outside this range has been incorrectly geocoded. We will remove these rows since there is no point in correcting since these will just be an issue later in the application when we try to display these suburbs.

No coordinates have been geocoded too 'low' (i.e., less than -39.2), but the north border of Victoria is roughly divided into 5 linear segments of a piece-wise function. Luckily, no coordinates are above this line but also within the longitude range specified above (i.e., no action needs to be taken).

bad_localities = pd.concat([new_suburb_data.pipe(lambda x: x[x["coordinates"].str[1] < 140.9]).sort_values("coordinates"), new_suburb_data.pipe(lambda x: x[x["coordinates"].str[1] > 150]).sort_values("coordinates")], ignore_index=True).locality
new_suburb_data = new_suburb_data[~new_suburb_data.locality.isin(bad_localities)]

new_suburb_data.shape
new_suburb_data.head()

## Save to new file

Now save the dataset to a new file for use in other files.

new_suburb_data.to_csv(os.path.join(DERIVED_DATA_PATH, "SuburbMetadata.csv"), index=False)
