==> notebooks/1-DataMerge+FeatureExtraction.ipynb <==
# Data Merge

This notebook outlines the steps taken to merge the required suburb attributes into a single dataset.

It then shows the steps taken in creating and populating the sqlalchemy database which is to be used in the app.

import os
import numpy as np
import pandas as pd

import helper

# make all output interactive
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# Define universal constants
# ================================
DATA_PATH = os.path.join(os.getcwd(), "../data/")
RAW_DATA_PATH = os.path.join(DATA_PATH, "raw")
DERIVED_DATA_PATH = os.path.join(DATA_PATH, "derived")
CORR_DATA_PATH = os.path.join(DATA_PATH, "correspondence")

## Processing

Each 'suburb' in the dataset, and hence the application, will be based on the Suburb and Locality non-ABS structure.


# Define merge-section-specific constants
# ================================

# 2016 census data
DATA_2016_CENSUS_DATA_PATH = os.path.join(RAW_DATA_PATH, "2016_GCP_SSC_for_VIC_short-header/2016 Census GCP State Suburbs for VIC")
DATA_2016_CENSUS_METADATA_PATH = os.path.join(RAW_DATA_PATH, "2016_GCP_SSC_for_VIC_short-header/Metadata")
DATA_2016_CENSUS_DATA_FILENAMES = os.listdir(DATA_2016_CENSUS_DATA_PATH)

# 2021 census data
DATA_2021_CENSUS_DATA_PATH = os.path.join(RAW_DATA_PATH, "2021_GCP_SAL_for_VIC_short-header/2021 Census GCP Suburbs and Localities for VIC")
DATA_2021_CENSUS_METADATA_PATH = os.path.join(RAW_DATA_PATH, "2021_GCP_SAL_for_VIC_short-header/Metadata")
DATA_2021_CENSUS_DATA_FILENAMES = os.listdir(DATA_2021_CENSUS_DATA_PATH)

# 2021 crime statistics data
DATA_2021_CRIME_DATA_PATH = os.path.join(RAW_DATA_PATH, "Data_Tables_LGA_Criminal_Incidents_Year_Ending_December_2021.xlsx")

# Conversion between 2016 SSC codes and 2021 SAL codes
DATA_CONVERT_SSC_SAL_PATH = os.path.join(RAW_DATA_PATH, "convert-SAL-2016-2021.csv")

# All meta data about every suburb/locality
LOCALITIES_DETAILS_PATH = os.path.join(RAW_DATA_PATH, "australian_postcodes.csv")


### Prepare census dataset (2021 only for now)

# All localities meta data
# ========================

suburb_metadata_df = helper.getSuburbsMetadata()
suburb_metadata_df

# 2021 census meta data
# =====================

# read file
data_2021_census_metadata_df = pd.read_excel(
    os.path.join(DATA_2021_CENSUS_METADATA_PATH, "2021Census_geog_desc_1st_release.xlsx"),
    sheet_name="2021_ASGS_Non_ABS_Structures"
)
# filter for victorian regions
data_2021_census_metadata_df = data_2021_census_metadata_df[data_2021_census_metadata_df["Census_Code_2021"].str.contains("^SAL2")]
# split lga and state component from name column where multiple places have the same name
data_2021_census_metadata_df["Census_Name_2021"] = data_2021_census_metadata_df["Census_Name_2021"].str.replace("\s*\(Vic.\)$", "", regex=True)
data_2021_census_metadata_df["Census_Name_2021"] = data_2021_census_metadata_df["Census_Name_2021"].str.replace("^(.*?)\s*\((.*?) - Vic.\)$", lambda m: f"{m.group(1)}, {m.group(2)}", regex=True)
data_2021_census_metadata_df[["locality", "lgaregion"]] = data_2021_census_metadata_df["Census_Name_2021"].str.split(", ", n=1, expand=True)
# split dataset horizontally between those with lga data and those without
data_2021_census_metadata_df_lga = data_2021_census_metadata_df[data_2021_census_metadata_df["lgaregion"].notnull()]
data_2021_census_metadata_df_nolga = data_2021_census_metadata_df[data_2021_census_metadata_df["lgaregion"].isnull()]

# TODO: Handle when the name is duplicate and the LGA region is specified
data_2021_census_metadata_df_nolga

# Combine all locality details and 2021 census meta data

suburb_metadata_df = suburb_metadata_df.merge(
    data_2021_census_metadata_df_nolga,
    how="right",  # only keep the rows which have data in the 2021 census
    on="locality",
    suffixes=[None, "_census"]
)
suburb_metadata_df

### Prepare crime dataset

# read file
crime_df = pd.read_excel(DATA_2021_CRIME_DATA_PATH, sheet_name="Table 03")


# filter for 2021 statistics
crime_df = crime_df[crime_df["Year"] == 2021]


# drop unnecessary columns
crime_df = crime_df.drop(["Year", "Year ending", "Postcode", "Unnamed: 9"], axis=1)


# group by locality and offence division and pivot table
crime_df = crime_df.groupby(["Suburb/Town Name", "Offence Division"], ).sum().unstack()


# clean up groupby result
crime_df = crime_df.reset_index()
crime_df.columns = ["locality", "A Crimes against the person", "B Property and deception offences",
                "C Drug offences", "D Public order and security offences", "E Justice procedures offences", "F Other offences"]


# remove null values with 0
crime_df.fillna(0, inplace=True)

crime_df

### Prepare environment and industry data

# Environment dataset
# ===================

env_df = pd.read_excel(
    os.path.join(RAW_DATA_PATH, "2011-2020 ABS Regional Data/LandEnvironment.xlsx"),
    sheet_name="Table 1",
    header=6,
    skipfooter=6
)
env_df = env_df[env_df["Code"].astype(str).str.contains("^2[0-9]{8}")]
env_df = env_df[env_df["Year"] == 2020]
env_df = env_df.replace("-", np.nan)
env_df = env_df.fillna(0)

env_df


# Industry dataset
# ===================

ind_df = pd.read_excel(
    os.path.join(RAW_DATA_PATH, "2011-2020 ABS Regional Data/EconomyIndustry.xlsx"),
    sheet_name="Table 1",
    header=6,
    skipfooter=6
)
ind_df = ind_df[ind_df["Code"].astype(str).str.contains("^2[0-9]{8}")]
ind_df = ind_df[ind_df["Year"] == 2020]
ind_df = ind_df.replace("-", np.nan)
ind_df = ind_df.fillna(0)

ind_df


### Merge datasets

- 2016 census data uses SSC
- 2021 census data uses SAL
- 2021 crime data uses suburb name
- Environment data (2020) uses SA2 (2016)
- Industry data (2020) uses SA2 (2016)

data = suburb_metadata_df.copy()

# Join data from 2021 Census
# ==========================

for file in DATA_2021_CENSUS_DATA_FILENAMES:
    file_path = os.path.join(DATA_2021_CENSUS_DATA_PATH, file)
    temp = pd.read_csv(file_path)
    data = data.merge(
        temp, 
        how="left", 
        left_on="Census_Code_2021", 
        right_on="SAL_CODE_2021",
        suffixes=[None, "_" + file.split("_")[1]]
    )


# Join data from crime statistics dataset
# =======================================

data = data.merge(crime_df, how="left", on="locality")


# Join data from environment and industry datasets
# ================================================

data = data.merge(env_df, how="left", left_on="sa2_code_2016", right_on="Code")
data = data.merge(ind_df, how="left", left_on="sa2_code_2016", right_on="Code")


data.head()

data.to_csv("../data/derived/SuburbRawdata.csv", index=False)

## Feature extraction

new_data = suburb_metadata_df.copy()

### Accomodation attributes

# acc_house_onebed
new_data["acc_house_onebed"] = (data["Separate_house_NofB_1"] + data["Se_d_r_or_t_h_t_Tot_NofB_1"])/data['Tot_Tot_P']
new_data["acc_house_onebed"].fillna(0, inplace=True)
# acc_house_twobed
new_data["acc_house_twobed"] = (data["Separate_house_NofB_2"] + data["Se_d_r_or_t_h_t_Tot_NofB_2"])/data['Tot_Tot_P']
new_data["acc_house_twobed"].fillna(0, inplace=True)
# acc_house_threebed
new_data["acc_house_threebed"] = (data["Separate_house_NofB_3"] + data["Se_d_r_or_t_h_t_Tot_NofB_3"])/data['Tot_Tot_P']
new_data["acc_house_threebed"].fillna(0, inplace=True)
# acc_house_fourplusbed
new_data["acc_house_fourplusbed"] = (data["Separate_house_NofB_4"] + data["Se_d_r_or_t_h_t_Tot_NofB_4"] + data["Separate_house_NofB_5"] + data["Se_d_r_or_t_h_t_Tot_NofB_5"] + data["Separate_house_NofB_6_or_m"] + data["Se_d_r_or_t_h_t_Tot_NofB_6_m"])/data['Tot_Tot_P']
new_data["acc_house_fourplusbed"].fillna(0, inplace=True)
# acc_apartment_onebed
new_data["acc_apartment_onebed"] = data["Flt_apart_Tot_NofB_1"]/data['Tot_Tot_P']
new_data["acc_apartment_onebed"].fillna(0, inplace=True)
# acc_apartment_twobed
new_data["acc_apartment_twobed"] = data["Flt_apart_Tot_NofB_2"]/data['Tot_Tot_P']
new_data["acc_apartment_twobed"].fillna(0, inplace=True)
# acc_apartment_threebed
new_data["acc_apartment_threebed"] = data["Flt_apart_Tot_NofB_3"]/data['Tot_Tot_P']
new_data["acc_apartment_threebed"].fillna(0, inplace=True)
# acc_apartment_fourplusbed
new_data["acc_apartment_fourplusbed"] = (data["Flt_apart_Tot_NofB_4"] + data["Flt_apart_Tot_NofB_5"] + data["Flt_apart_Tot_NofB_6_or_m"])/data['Tot_Tot_P']
new_data["acc_apartment_fourplusbed"].fillna(0, inplace=True)
# acc_rented_house_relative
new_data["acc_rented_house_relative"] = (data["R_Tot_DS_Sep_house"] + data["R_Tot_DS_SemiD_ro_or_tce_h_th"]) / (data["Total_DS_Sep_house"] + data["Total_DS_SemiD_ro_or_tce_h_th"])
new_data["acc_rented_house_relative"].fillna(0, inplace=True)
# acc_rented_apartment_relative
new_data["acc_rented_apartment_relative"] = data["R_Tot_DS_Flat_apart"] / data["Total_DS_Flat_apart"]
new_data["acc_rented_apartment_relative"].fillna(0, inplace=True)
# acc_shared_relative
new_data["acc_shared_relative"] = data["Tot_Group_H"] / data["Tot_Tot_G42"]
new_data["acc_shared_relative"].fillna(0, inplace=True)
# acc_rent_1_74
new_data["acc_rent_1_74"] = data["R_1_74_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_1_74"].fillna(0, inplace=True)
# acc_rent_75_99
new_data["acc_rent_75_99"] = data["R_75_99_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_75_99"].fillna(0, inplace=True)
# acc_rent_100_149
new_data["acc_rent_100_149"] = data["R_100_149_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_100_149"].fillna(0, inplace=True)
# acc_rent_150_199
new_data["acc_rent_150_199"] = data["R_150_199_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_150_199"].fillna(0, inplace=True)
# acc_rent_200_224
new_data["acc_rent_200_224"] = data["R_200_224_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_200_224"].fillna(0, inplace=True)
# acc_rent_225_274
new_data["acc_rent_225_274"] = data["R_225_274_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_225_274"].fillna(0, inplace=True)
# acc_rent_275_349
new_data["acc_rent_275_349"] = data["R_275_349_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_275_349"].fillna(0, inplace=True)
# acc_rent_350_449
new_data["acc_rent_350_449"] = data["R_350_449_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_350_449"].fillna(0, inplace=True)
# acc_rent_450_549
new_data["acc_rent_450_549"] = data["R_450_549_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_450_549"].fillna(0, inplace=True)
# acc_rent_550_649
new_data["acc_rent_550_649"] = data["R_550_649_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_550_649"].fillna(0, inplace=True)
# acc_rent_650_749
new_data["acc_rent_650_749"] = data["R_650_749_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_650_749"].fillna(0, inplace=True)
# acc_rent_750_849
new_data["acc_rent_750_849"] = data["R_750_849_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_750_849"].fillna(0, inplace=True)
# acc_rent_850_949
new_data["acc_rent_850_949"] = data["R_850_949_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_850_949"].fillna(0, inplace=True)
# acc_rent_950_plus
new_data["acc_rent_950_plus"] = data["R_950_over_Tot"]/data['Tot_Tot_P']
new_data["acc_rent_950_plus"].fillna(0, inplace=True)

### Industry & Environment attributes

new_data["env_retail"] = data["Retail trade (no.)"]/data['Tot_Tot_P']
new_data["env_retail"].fillna(0, inplace=True)

new_data["env_accomodation_food"] = data["Accommodation and food services (no.)"]/data['Tot_Tot_P']
new_data["env_accomodation_food"].fillna(0, inplace=True)

new_data["env_public_admin"] = data["Public administration and safety (no.)"]/data['Tot_Tot_P']
new_data["env_public_admin"].fillna(0, inplace=True)

new_data["env_healthcare_social_assist"] = data["Health care and social assistance (no.)"]/data['Tot_Tot_P']
new_data["env_healthcare_social_assist"].fillna(0, inplace=True)

new_data["env_arts_recreation"] = data["Arts and recreation services (no.)"]/data['Tot_Tot_P']
new_data["env_arts_recreation"].fillna(0, inplace=True)

new_data["env_rental_hiring_realestate"] = data["Rental, hiring and real estate services (no.)"]/data['Tot_Tot_P']
new_data["env_rental_hiring_realestate"].fillna(0, inplace=True)

new_data["env_parks"] = data["Protected land area (%)"]/data['Tot_Tot_P']
new_data["env_parks"].fillna(0, inplace=True)

### Demographics attributes

new_data["dem_students_relative"] = data["Tert_Tot_Tert_P"] / data["Tot_Tot_P"]
new_data["dem_students_relative"].fillna(0, inplace=True)


### Transport attributes

The code used to determine these features is located in the [1.2-TransportAttributes.ipynb](1.2-TransportAttributes.ipynb) jupyter notebook.

# read saved scores from disk
suburb_transport_score_df = pd.read_csv(os.path.join(DERIVED_DATA_PATH, "SuburbTransportScores.csv"))
# add new columns based on `locality` column
new_data = new_data.merge(suburb_transport_score_df)


### Safety attributes

# Offence division A
new_data["saf_crime_person"] = data["A Crimes against the person"] / data["Tot_Tot_P"]
new_data["saf_crime_person"].fillna(0, inplace=True)
# Offence division B
new_data["saf_crime_property"] = data["B Property and deception offences"] / data["Tot_Tot_P"]
new_data["saf_crime_property"].fillna(0, inplace=True)
# Offence division C
new_data["saf_drug_offences"] = data["C Drug offences"] / data["Tot_Tot_P"]
new_data["saf_drug_offences"].fillna(0, inplace=True)
# Offence division D
new_data["saf_order_security"] = data["D Public order and security offences"] / data["Tot_Tot_P"]
new_data["saf_order_security"].fillna(0, inplace=True)
# Offence division E
new_data["saf_justice_procedure"] = data["E Justice procedures offences"] / data["Tot_Tot_P"]
new_data["saf_justice_procedure"].fillna(0, inplace=True)
# Offence division F
new_data["saf_other"] = data["F Other offences"] / data["Tot_Tot_P"]
new_data["saf_other"].fillna(0, inplace=True)

### Commute attributes

The code used to determine these features are located in the [1.1-CommuteAttributes.ipynb](1.1-CommuteAttributes.ipynb) jupyter notebook.

# read dataframe creating in other juypter notebook
suburb_campus_dist_df = pd.read_csv(os.path.join(DERIVED_DATA_PATH, "SuburbCampusDist.csv"))
# suburb_campus_dist_df.drop("id", axis=1, inplace=True)

# merge dataframes
new_data = new_data.merge(suburb_campus_dist_df, on="locality")


### Clean and export

export_df = new_data.drop(["lgaregion", "sa1_code_2016", "sa1_code_2021", "sa2_code_2016", "sa2_name_2016", "sa2_code_2021", "sa2_name_2021", "sa3_code_2016", "sa3_name_2016", "sa3_code_2021", "sa3_name_2021",
                          "sa4_code_2016", "sa4_name_2016", "sa4_code_2021", "sa4_name_2021", "ASGS_Structure", "Census_Code_2021", "AGSS_Code_2021", "Census_Name_2021", "Area sqkm", "lgaregion_census", ], axis=1)
export_df.replace(np.inf, 0, inplace=True)

export_df.to_csv("../data/derived/SuburbData.csv", index=False)
