==> notebooks/1.1-CommuteAttributes.ipynb <==
# University-Suburb distance calculation

import os
import pandas as pd
from geopy.geocoders import GoogleV3  # for determining uni campus coordinates
from geopy.distance import geodesic  # for calculating distance between coordinates

import helper

# make all output interactive
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"


# Define universal constants
# ================================
DATA_PATH = os.path.join(os.getcwd(), "../data/")
RAW_DATA_PATH = os.path.join(DATA_PATH, "raw")
DERIVED_DATA_PATH = os.path.join(DATA_PATH, "derived")
CORR_DATA_PATH = os.path.join(DATA_PATH, "correspondence")


geolocator = GoogleV3(api_key=os.getenv("GMAPS_APIKEY"))

In order for the above `geolocator` code to work, VSCode must be opened from the terminal using the following code to ensure the environment variable for the Google Maps API key is set correctly and inherited by the session:

``` bash
> source ~/.bash-profile
> code .
```

CAMPUS_NAMES = [
    "Swinbourne University of Technology Hawthorn Campus",
    "Swinbourne University of Technology Croydon Campus",
    "Swinbourne University of Technology Wantirna Campus",
    "Deakin University Burwood Campus",
    "Deakin University Geelong Campus",
    "Deakin University Warrnambool Campus",
    "Federation University Australia Ballarat Campus",
    "Federation University Australia Churchill Campus",
    "Federation University Australia Berwick Campus",
    "Federation University Australia Wimmera Campus",
    "La Trobe University Melbourne Campus",
    "La Trobe University Bendigo Campus",
    "La Trobe University Shepparton Campus",
    "La Trobe University Wodonga Campus",
    "La Trobe University Mildura Campus",
    "Monash University Clayton Campus",
    "Monash University Caulfield Campus",
    "Monash University Peninsula Campus",
    "Monash University Parkville Campus",
    "RMIT University Melbourne Campus",
    "Swinburne University of Technology Hawthorne Campus",
    "Swinburne University of Technology Croydon Campus",
    "Swinburne University of Technology Wantirna Campus",
    "The University of Melbourne Parkville Campus",
    "The University of Melbourne Southbank Campus",
    "The University of Melbourne Burnley Campus",
    "The University of Melbourne Dookie Campus",
    "The University of Melbourne Creswick Campus",
    "The University of Melbourne Werribee Campus",
    "The University of Melbourne Shepparton Campus",
    "Victoria University Melbourne Campus",
    "Victoria University Footscray Campus",
    "Victoria University St Albans Campus",
    "Victoria University Sunshine Campus",
    "Victoria University Werribee Campus",
    "Australian Catholic University Ballarat Campus",
    "Australian Catholic University Melbourne Campus",
    "Torrens University Australia Melbourne Campus",
]

CAMPUSES_CODES = [
    "com_swinbourne_hawthorn",
    "com_swinbourne_croydon",
    "com_swinbourne_wantirna",
    "com_deakin_burwood",
    "com_deakin_geelong",
    "com_deakin_warrnambool",
    "com_federation_ballarat",
    "com_federation_churchill",
    "com_federation_berwick",
    "com_federation_wimmera",
    "com_latrobe_melbourne",
    "com_latrobe_bendigo",
    "com_latrobe_shepparton",
    "com_latrobe_wodonga",
    "com_latrobe_mildura",
    "com_monash_clayton",
    "com_monash_caulfield",
    "com_monash_peninsula",
    "com_monash_parkville",
    "com_rmit_melbourne",
    "com_swinburne_hawthorne",
    "com_swinburne_croydon",
    "com_swinburne_wantirna",
    "com_unimelb_parkville",
    "com_unimelb_southbank",
    "com_unimelb_burnley",
    "com_unimelb_dookie",
    "com_unimelb_creswick",
    "com_unimelb_werribee",
    "com_unimelb_shepparton",
    "com_vicuni_melbourne",
    "com_vicuni_footscray",
    "com_vicuni_stalbans",
    "com_vicuni_sunshine",
    "com_vicuni_werribee",
    "com_catholic_ballarat",
    "com_catholic_melbourne",
    "com_torrens_melbourne",
]


## Get campus coordinates

campus_coordinates = []

for name in CAMPUS_NAMES:
    point = geolocator.geocode(name).point
    tup = (point.latitude, point.longitude)
    campus_coordinates.append(tup)

campus_coordinates


## Get distance from campus to suburb

We will be using the standard derived metadata file for the suburb coordinates, found in [`SuburbMetadata.csv`](../data/derived/SuburbMetadata.csv).

# All localities meta data
# ========================

# get data
suburb_metadata_df = helper.getSuburbsMetadata()
# convert to numpy array for faster operations
suburb_metadata = suburb_metadata_df[["locality", "coordinates"]].to_numpy()
suburb_metadata


dist_matrix = []

for details in suburb_metadata:
    # prepare variables for locality details from list
    suburb_name = details[0]
    suburb_coordinates = details[1]

    for campus_code, campus_coord in zip(CAMPUSES_CODES, campus_coordinates):
        dist = geodesic(suburb_coordinates, campus_coord).km
        dist_matrix.append([
            suburb_name,
            campus_code,
            dist
        ])

dist_matrix

## Cleaning

Round numbers and pivot for column attributes for each campus.

dist_df = pd.DataFrame(dist_matrix, columns=["locality", "campus", "distance_km"])

# round distances to nearest km
dist_df["distance_km"] = dist_df["distance_km"].round(0).astype(int)

# pivot table
dist_df = dist_df\
    .pivot(index="locality", columns="campus", values="distance_km")\
    .reset_index()

dist_df.shape
dist_df.head()

## Save

dist_df.to_csv(os.path.join(DERIVED_DATA_PATH, "SuburbCampusDist.csv"), index=False)
