==> notebooks/3-GenerateTrainingData.ipynb <==
# Generate training dataset

import os
import numpy as np
import pandas as pd
from sklearn.preprocessing import MinMaxScaler

# make all output interactive
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import helper

# Define file structure constants
# ================================
DATA_PATH = os.path.join(os.getcwd(), "../data/")
RAW_DATA_PATH = os.path.join(DATA_PATH, "raw")
DERIVED_DATA_PATH = os.path.join(DATA_PATH, "derived")

# Read datasets
# ================================
# main suburb data
suburb_data = helper.getSuburbData()
# generated preference data
preference_data = helper.getPreferenceData()
# suburb campus distance matrix
suburb_campus_dist = pd.read_csv("../data/derived/SuburbCampusDist.csv")
# suburb raw data
suburb_rawdata = pd.read_csv("../data/derived/SuburbRawdata.csv")

# All the predictor variables in training data
attributes = ['acc_house_onebed', 'acc_house_twobed', 'acc_house_threebed', 'acc_house_fourplusbed', 'acc_apartment_onebed', 'acc_apartment_twobed', 'acc_apartment_threebed', 'acc_apartment_fourplusbed', 'acc_rented_house_relative', 'acc_rented_apartment_relative', 'acc_shared_relative', 'acc_rent_1_74', 'acc_rent_75_99', 'acc_rent_100_149', 'acc_rent_150_199', 'acc_rent_200_224', 'acc_rent_225_274', 'acc_rent_275_349', 'acc_rent_350_449', 'acc_rent_450_549', 'acc_rent_550_649', 'acc_rent_650_749', 'acc_rent_750_849', 'acc_rent_850_949', 'acc_rent_950_plus', 'env_retail', 'env_accomodation_food', 'env_public_admin', 'env_healthcare_social_assist', 'env_arts_recreation', 'env_rental_hiring_realestate', 'env_parks', 'dem_students_relative', 'tra_train', 'tra_bus', 'tra_tram', 'saf_crime_person', 'saf_crime_property', 'saf_drug_offences', 'saf_order_security', 'saf_justice_procedure', 'saf_other', 'com_max_dist', 'com_catholic_ballarat', 'com_catholic_melbourne', 'com_deakin_burwood', 'com_deakin_geelong', 'com_deakin_warrnambool', 'com_federation_ballarat', 'com_federation_berwick', 'com_federation_churchill', 'com_federation_wimmera', 'com_latrobe_bendigo', 'com_latrobe_melbourne', 'com_latrobe_mildura', 'com_latrobe_shepparton', 'com_latrobe_wodonga', 'com_monash_caulfield', 'com_monash_clayton', 'com_monash_parkville', 'com_monash_peninsula', 'com_rmit_melbourne', 'com_swinbourne_croydon', 'com_swinbourne_hawthorn', 'com_swinbourne_wantirna', 'com_swinburne_croydon', 'com_swinburne_hawthorne', 'com_swinburne_wantirna', 'com_torrens_melbourne', 'com_unimelb_burnley', 'com_unimelb_creswick', 'com_unimelb_dookie', 'com_unimelb_parkville', 'com_unimelb_shepparton', 'com_unimelb_southbank', 'com_unimelb_werribee', 'com_vicuni_footscray', 'com_vicuni_melbourne', 'com_vicuni_stalbans', 'com_vicuni_sunshine', 'com_vicuni_werribee']

## Normalisation

The values/scores for each 'quality' of the suburbs must be normalised between all suburbs to ensure that one attribute/'quality' does not dominate others. For example, an attribute such as `dem_students_relative` is very small (due to the relative nature of the attribute) but another attribute such as `com_monash_clayton` can be very large (since this attribute is determined by the distance in kilometers from the campus to the suburb).

Normalisation is done using the following equation: $$X_{normalised} = \frac{X - min(X)}{max(X) - min(X)}$$

def normalise_column(column):
    return (column - column.min()) / (column.max() - column.min())

def normalise_dataframe(df: pd.DataFrame, columns: list = None, inplace: bool = False) -> pd.DataFrame:
    # use all columns by default (columns = None)
    if columns is None:
        columns = df.columns
    # link or copy depending on inplace method
    new_df = df if inplace is True else df.copy()
    # iterate over columns and replace/create column with normalised
    for col in columns:
        new_df[col] = normalise_column(new_df[col])
    return new_df

normalised_suburb_data = normalise_dataframe(
    suburb_data, 
    list(suburb_data.columns[3:])
)
normalised_suburb_data.head()

## Inversion

Some of the suburb attributes must be inverted due to differing definitions of 'goodness'. For example, the attributes for commute themed qualities are 'worse' for a suburb when the value is large (i.e., a large distance is not desired), whereas for another attribute such as `dem_students_relative`, having a large value is good since the user is looking to have a higher density of students.

The attributes which must be inverted are the attributes from the commute theme and the safety theme (since the safety values are determined as high values = more crimes).

Since the attributes have been normalised in the previous section of the notebook, we can simply subtract the value within these columns from 1. This is possible since all values are now from 0 to 1 (inclusive). Any previously high values (which we want to become small), will do so when subtracted from 1. Likewise for the case of low values (which we want to become large).

def invert_dataframe(df: pd.DataFrame, columns: list = None, inplace: bool = False) -> pd.DataFrame:
    # use all columns by default (columns = None)
    if columns is None:
        columns = df.columns
    # link or copy depending on inplace method
    new_df = df if inplace is True else df.copy()
    # iterate over columns and replace/create column with inverted
    for col in columns:
        new_df[col] = 1 / new_df[col]
    return new_df

invertable_columns = [i for i in list(normalised_suburb_data.columns) if i.split("_", maxsplit=1)[0] in ["saf", "com"]]

inv_norm_suburb_data = invert_dataframe(normalised_suburb_data, invertable_columns)

inv_norm_suburb_data.head()

inv_norm_suburb_data.to_csv(os.path.join(DERIVED_DATA_PATH, "NormalizedSuburbData.csv"), index=False)

## Additional suburb data processing

We are going to remove all suburbs who don't have a university campus within 35 km. Remember, the preferences only allow the user to put a max distance of 30 km.

only_commute = suburb_data[suburb_data.columns[suburb_data.columns.str.contains("^com_", regex=True)]]

only_commute_dist_mask = only_commute.apply(lambda row: min(row) <= 35, axis=1)

new_suburb_data = suburb_data.where(only_commute_dist_mask, axis=0).dropna()

new_suburb_data

Now we remove any suburbs whose total renting population (`Tot_Tot_G40` in Raw Dataset) is less than 100 (since it is unlikely that a user would be able to find rented accomodation in this region) and where the population is less than 100 (since these are likely too small for comfortable living).

suburb_rawdata["rentable_per_capita"] = ((suburb_rawdata["Tot_Tot_G40"] / suburb_rawdata["Tot_P_P"]) * suburb_rawdata.Tot_P_P.mean()).fillna(0)

pop_locals = suburb_rawdata.query("rentable_per_capita > 100 and Tot_Tot_G40 > 100").locality

Combine these two filters together.

new_suburb_data = new_suburb_data.merge(pop_locals)

## Generate output suburb

Using `new_suburb_data` with the smaller refined list of suburbs, we now determine which suburb is best for each instance of generated preference data in `preference_data`.

- Ensure that the distance to the selected campus is within the `com_max_dist` input.
- Remove the bottom 10% of suburbs for selected transport options.

def get_campus(pref_row):
    temp_com = pref_row.where(pref_row[pref_row.index[pref_row.index.str.contains("^com_", regex=True)]] > 0).dropna().index
    if len(temp_com) > 1:
        return temp_com[1]
    else:
        return None

def get_suburb(pref_row):
    # copy suburb dataframe so that we can make edits
    copy_suburb_data = new_suburb_data.copy()

    # filter by campus distance when that is required (not None priority)
    if (campus := get_campus(pref_row)) is not None:
        # ensure the user max distance is not smaller than the closest suburb
        if (minimum := min(copy_suburb_data[get_campus(pref_row)])) > pref_row.com_max_dist:
            pref_row.com_max_dist = minimum * 10

        copy_suburb_data = copy_suburb_data.query(f"{get_campus(pref_row)} <= {pref_row.com_max_dist}")


    # invert and normalise the data
    copy_suburb_data = invert_dataframe(copy_suburb_data, invertable_columns)
    # copy_suburb_data.iloc[:,3:] = copy_suburb_data.iloc[:,3:] * 1000
    copy_suburb_data = copy_suburb_data.replace(np.inf, 0)
    copy_suburb_data = normalise_dataframe(copy_suburb_data, copy_suburb_data.columns[3:])

    # dot product each row of suburb dataset with pref_data and get the top 'score'
    best_suburb_index = copy_suburb_data.iloc[:,3:].multiply(pref_row).sum(axis=1).nlargest(1).index[0]
    # return the name of the locality which is 'best'
    return copy_suburb_data.locality[best_suburb_index]

training_data = preference_data.assign(suburb=(
    preference_data.apply(lambda row: get_suburb(row), axis=1)
))

training_data.suburb.value_counts()

training_data.to_csv(os.path.join(DERIVED_DATA_PATH, "TrainingData.csv"), index=False)
