==> notebooks/4-CreateDecisionTree.ipynb <==
# Create Decision Tree Model

import os
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
import joblib


# make all output interactive
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import helper

# Define file structure constants
# ================================
DATA_PATH = os.path.join(os.getcwd(), "../data/")
RAW_DATA_PATH = os.path.join(DATA_PATH, "raw")
DERIVED_DATA_PATH = os.path.join(DATA_PATH, "derived")

# Read datasets
# ================================
# training_data
training_data = helper.getTrainingData()

## Decision Tree Model

# X -> features, y -> label
X = training_data.loc[:, training_data.columns != 'suburb']
y = training_data['suburb']
  
# Dividing X, y into train and test data
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 0)
  
# Training a DecisionTreeClassifier
dtree_model = DecisionTreeClassifier()

dtree_model.fit(X_train, y_train)
# Save the model
joblib.dump(dtree_model, '../src/dtree_model.pkl')

## Predict on test set (REMOVE LATER)

dtree_predictions = dtree_model.predict(X_test)
  
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix
print(accuracy_score(y_test, dtree_predictions))

## Predict on new user input preference

# Load the decision tree model
dtree_model = joblib.load('../src/dtree_model.pkl')

# Example new user preference data
input_data = training_data.sample(1)
# Ensure order of columns in user preference data is same as training data
input_data = input_data[dtree_model.feature_names_in_]

# Predicted suburb with highest probability
predicted_suburb = dtree_model.predict(input_data)
print("Recommended Suburb:", predicted_suburb[0])

# List of predicted suburbs ranked by probability
predicted_classes = dtree_model.classes_
predicted_probabilities = dtree_model.predict_proba(input_data)

recommended_suburb_list = pd.DataFrame(list(zip(predicted_classes, predicted_probabilities[0])), columns=['suburb','probability']).sort_values(by='probability', ascending=False)
print(recommended_suburb_list)
